# apps/aicf-api/openapi.yaml
openapi: 3.1.0
info:
  title: AICF API
  version: 0.1.0
  description: ANM-native decentralized AI compute cloud API for aicf.animica.org.
servers:
  - url: https://api.aicf.animica.org
  - url: http://127.0.0.1:8099
paths:
...
  /internal/scheduler/tick:
    post:
      summary: Trigger scheduler assignment tick
      security:
        - InternalSecret: []
  /internal/chain-events/watch:
    get:
      summary: Poll chain-derived contract events
      security:
        - InternalSecret: []
  /internal/contract-jobs/scheduler/tick:
    post:
      summary: Schedule contract jobs to providers
      security:
        - InternalSecret: []
  /internal/contract-jobs/result-submitter/tick:
    post:
      summary: Promote commitment-only jobs when result references are available
      security:
        - InternalSecret: []
  /internal/contract-jobs/finalization/tick:
    post:
      summary: Finalize eligible contract jobs after challenge windows
      security:
        - InternalSecret: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey or SessionToken
    InternalSecret:
      type: apiKey
      in: header
      name: x-internal-secret